+++ hutool-core/src/main/java/cn/hutool/core/net/Ipv4Util.java
@@ -3,13 +3,14 @@ package cn.hutool.core.net;
 import cn.hutool.core.collection.ListUtil;
 import cn.hutool.core.convert.Convert;
 import cn.hutool.core.lang.Assert;
-import cn.hutool.core.lang.Validator;
+import cn.hutool.core.lang.PatternPool;
 import cn.hutool.core.util.CharUtil;
 import cn.hutool.core.util.StrUtil;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.regex.Matcher;
 
 /**
  * IPV4地址工具类
@@ -20,6 +21,7 @@ import java.util.Objects;
  * @since 5.4.1
  */
 public class Ipv4Util {
+
 	/**
 	 * IP段的分割符
 	 */
@@ -149,18 +151,25 @@ public class Ipv4Util {
 	/**
 	 * 根据ip地址(xxx.xxx.xxx.xxx)计算出long型的数据
 	 * 方法别名：inet_aton
+	 *
 	 * @param strIP IP V4 地址
 	 * @return long值
 	 */
 	public static long ipv4ToLong(String strIP) {
-		Validator.validateIpv4(strIP, "Invalid IPv4 address!");
-		final long[] ip = Convert.convert(long[].class, StrUtil.split(strIP, CharUtil.DOT));
-		return (ip[0] << 24) + (ip[1] << 16) + (ip[2] << 8) + ip[3];
+		final Matcher matcher = PatternPool.IPV4.matcher(strIP);
+		if (matcher.matches()) {
+			return matchAddress(matcher);
+		}
+//		Validator.validateIpv4(strIP, "Invalid IPv4 address!");
+//		final long[] ip = Convert.convert(long[].class, StrUtil.split(strIP, CharUtil.DOT));
+//		return (ip[0] << 24) + (ip[1] << 16) + (ip[2] << 8) + ip[3];
+		throw new IllegalArgumentException("Invalid IPv4 address!");
 	}
 
 	/**
 	 * 根据 ip/掩码位 计算IP段的起始IP（字符串型）
 	 * 方法别名：inet_ntoa
+	 *
 	 * @param ip      给定的IP，如218.240.38.69
 	 * @param maskBit 给定的掩码位，如30
 	 * @return 起始IP的字符串表示
@@ -195,9 +204,7 @@ public class Ipv4Util {
 	 * 根据子网掩码转换为掩码位
 	 *
 	 * @param mask 掩码的点分十进制表示，例如 255.255.255.0
-	 *
 	 * @return 掩码位，例如 24
-	 *
 	 * @throws IllegalArgumentException 子网掩码非法
 	 */
 	public static int getMaskBitByMask(String mask) {
@@ -282,7 +289,6 @@ public class Ipv4Util {
 	 * 判断掩码是否合法
 	 *
 	 * @param mask 掩码的点分十进制表示，例如 255.255.255.0
-	 *
 	 * @return true：掩码合法；false：掩码不合法
 	 */
 	public static boolean isMaskValid(String mask) {
@@ -293,7 +299,6 @@ public class Ipv4Util {
 	 * 判断掩码位是否合法
 	 *
 	 * @param maskBit 掩码位，例如 24
-	 *
 	 * @return true：掩码位合法；false：掩码位不合法
 	 */
 	public static boolean isMaskBitValid(int maskBit) {
@@ -314,6 +319,20 @@ public class Ipv4Util {
 	private static Long getEndIpLong(String ip, int maskBit) {
 		return getBeginIpLong(ip, maskBit)
 				+ ~ipv4ToLong(getMaskByMaskBit(maskBit));
+	}
+
+	/**
+	 * 将匹配到的Ipv4地址的4个分组分别处理
+	 *
+	 * @param matcher 匹配到的Ipv4正则
+	 * @return ipv4对应long
+	 */
+	private static long matchAddress(Matcher matcher) {
+		long addr = 0;
+		for (int i = 1; i <= 4; ++i) {
+			addr |= Long.parseLong(matcher.group(i)) << 8 * (4 - i);
+		}
+		return addr;
 	}
 	//-------------------------------------------------------------------------------- Private method end
 }