+++ hutool-core/src/test/java/cn/hutool/core/net/Ipv4UtilTest.java
@@ -1,11 +1,10 @@
 package cn.hutool.core.net;
 
-import cn.hutool.core.lang.Console;
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.function.ThrowingRunnable;
 
 import java.util.List;
-import org.junit.function.ThrowingRunnable;
 
 public class Ipv4UtilTest {
 
@@ -40,7 +39,7 @@ public class Ipv4UtilTest {
 		String ip = "192.168.1.1";
 		final int maskBitByMask = Ipv4Util.getMaskBitByMask("255.255.255.0");
 		final String endIpStr = Ipv4Util.getEndIpStr(ip, maskBitByMask);
-		Console.log(endIpStr);
+		Assert.assertEquals("192.168.1.255", endIpStr);
 	}
 
 	@Test
@@ -74,5 +73,17 @@ public class Ipv4UtilTest {
 	public void isMaskBitInvalidTest() {
 		boolean maskBitValid = Ipv4Util.isMaskBitValid(33);
 		Assert.assertFalse("掩码位非法检验", maskBitValid);
+	}
+
+	@Test
+	public void ipv4ToLongTest(){
+		long l = Ipv4Util.ipv4ToLong("127.0.0.1");
+		Assert.assertEquals(2130706433L, l);
+		l = Ipv4Util.ipv4ToLong("114.114.114.114");
+		Assert.assertEquals(1920103026L, l);
+		l = Ipv4Util.ipv4ToLong("0.0.0.0");
+		Assert.assertEquals(0L, l);
+		l = Ipv4Util.ipv4ToLong("255.255.255.255");
+		Assert.assertEquals(4294967295L, l);
 	}
 }