+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-# 5.7.16 (2021-10-28)
+# 5.7.16 (2021-10-31)
 
 ### 🐣新特性
 * 【core   】     增加DateTime.toLocalDateTime
@@ -15,6 +15,10 @@
 * 【extra  】     Sftp增加put和upload重载（issue#I4FGDH@Gitee）
 * 【core   】     TemporalUtil增加toChronoUnit、toTimeUnit方法（issue#I4FGDH@Gitee）
 * 【core   】     StopWatch增加prettyPrint重载（issue#1910@Github）
+* 【core   】     修改RegexPool中Ipv4正则
+* 【json   】     Filter改为MutablePair，以便编辑键值对（issue#1921@Github）
+* 【core   】     Opt增加peeks方法（pr#445@Gitee）
+* 【extra  】     MailAccount中user默认值改为邮箱全称（issue#I4FYVY@Gitee）
 
 ### 🐞Bug修复
 * 【core   】     修复UrlBuilder.addPath歧义问题（issue#1912@Github）
@@ -22,6 +26,7 @@
 * 【core   】     修复CharSequenceUtil.wrapIfMissing预定义长度计算问题（issue#I4FDZ2@Gitee）
 * 【poi    】     修复合并单元格为日期时，导出单元格数据为数字问题（issue#1911@Github）
 * 【core   】     修复CompilerUtil.getFileManager参数没有使用的问题（issue#I4FIO6@Gitee）
+* 【core   】     修复NetUtil.isInRange的cidr判断问题（pr#1917@Github）
 
 -------------------------------------------------------------------------------------------------------------
 
@@ -39,6 +44,8 @@
 * 【core   】     ZipUtil增加append方法（pr#441@Gitee）
 * 【core   】     CollUtil增加重载（issue#I4E9FS@Gitee）
 * 【core   】     CopyOptions新增setFieldValueEditor（issue#I4E08T@Gitee）
+* 【core   】     增加SystemPropsUtil（issue#1918@Gitee）
+* 【core   】     增加`hutool.date.lenient`系统属性（issue#1918@Gitee）
 
 ### 🐞Bug修复
 * 【core   】     修复CollUtil.isEqualList两个null返回错误问题（issue#1885@Github）